requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
opencv-python~=4.7.0.72
numpy~=1.25.0